Change account oauth feature to theme 2
