+++ feature/account/oauth/src/main/kotlin/app/k9mail/feature/account/oauth/ui/view/SignInWithGoogleButton.kt
@@ -22,10 +22,8 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
 import app.k9mail.feature.account.oauth.R
 import androidx.compose.material.Button as MaterialButton
 
@@ -121,26 +119,5 @@ private fun getTextColor(isLight: Boolean): Color {
         Color(0x87000000)
     } else {
         Color(0xFFFFFFFF)
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun SignInWithGoogleButtonPreview() {
-    PreviewWithThemes {
-        SignInWithGoogleButton(
-            onClick = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun SignInWithGoogleButtonDisabledPreview() {
-    PreviewWithThemes {
-        SignInWithGoogleButton(
-            onClick = {},
-            enabled = false,
-        )
     }
 }
